Fix compatibility with Pillow 13
Pillow allows you to save 32-bit tensors as PNG. However, PNG doesn't
support 32-bit images, so doing this actually converts the data to 16-bit.
Apparently Pillow 13 is going to stop supporting this implicit conversion,
so we have to do it ourselves.

This is technically a change in behavior, because Pillow handles
out-of-range values by clipping them to 0xffff, while Datumaro's `save_image`
just chops off the high bits (via `ndarray.astype`). I don't think it really
matters, though, since data is lost either way.

+++ src/datumaro/plugins/data_formats/kitti/exporter.py
@@ -1,5 +1,5 @@
 # Copyright (C) 2021 Intel Corporation
-# Copyright (C) 2022 CVAT.ai Corporation
+# Copyright (C) 2022-2026 CVAT.ai Corporation
 #
 # SPDX-License-Identifier: MIT
 
@@ -149,7 +149,7 @@ class KittiExporter(Exporter):
                         osp.join(self._save_dir, labelids_mask_path),
                         compiled_class_mask.class_mask,
                         apply_colormap=False,
-                        dtype=np.int32,
+                        dtype=np.uint16,
                     )
 
                     # TODO: optimize second merging
@@ -167,7 +167,7 @@ class KittiExporter(Exporter):
                         osp.join(self._save_dir, inst_path),
                         compiled_instance_mask.class_mask,
                         apply_colormap=False,
-                        dtype=np.int32,
+                        dtype=np.uint16,
                     )
 
                 bboxes = [a for a in item.annotations if a.type == AnnotationType.bbox]
